refactor: 14주차 피드백 일부 반영 및 folderFage 오류 수정
