+++ pages/index.tsx
@@ -1,16 +1,7 @@
-// import Link from "next/link";
-// import Gnb from "@/src/components/Gnb";
-// import Footer from "@/src/components/Footer";
-
 export default function Home() {
   return (
     <>
-      {/* <Gnb /> */}
       <h1>🌝HomePage🌝</h1>
-      {/* <Link href={"/shared"}>Shared</Link>
-      <br />
-      <Link href={"/folder"}>folder</Link> */}
-      {/* <Footer /> */}
     </>
   );
 }
