+++ src/components/AddLinkInput/AddLinkInput.module.scss
@@ -21,7 +21,7 @@
     border-radius: 15px;
     border: 1px solid #6d6afe;
     background: #fff;
-    background-image: url(../../assets/link.svg);
+    background-image: url(/assets/link.svg);
     background-repeat: no-repeat;
     background-position: 20px center;
     text-indent: 32px;
