+++ src/constant/constant.ts
@@ -1,0 +1,22 @@
+export const SNS_LIST = [
+  {
+    name: "Facebook",
+    href: "https://www.facebook.com/",
+    src: "/assets/facebook.svg",
+  },
+  {
+    name: "Twitter",
+    href: "https://www.twitter.com/",
+    src: "/assets/twitter.svg",
+  },
+  {
+    name: "Youtube",
+    href: "https://www.youtube.com/",
+    src: "/assets/youtube.svg",
+  },
+  {
+    name: "Instagram",
+    href: "https://www.instagram.com/",
+    src: "/assets/instagram.svg",
+  },
+];
